Mypy Compatible tests for cycler
- Convert to package instead of single file module
- add py.typed (including setup.py changes)
- add a CI workflow to run mypy
- Add the actual type hints inline

Type hints are generic over both the key type (most commonly strings)
and the value type.

I initially tried to be extra cute and encompass the type expansion that
can happen when cyclers of different types are added/multiplied, but
that proved to be more of a hassle to deal with than I was willing to
deal with initially.

As such, the example from the `concat` docstring will actually fail to
properly type check with implicit types (as Cycler[str, int] cannot be
concat'd with Cycler[str, str] and properly typecheck)

This may be solvable (resulting in a Cycler[str, int|str]) but that is
not yet implemented.
It may be resolved in downstream code somewhat crudely by typing both
operands explicitly as Cycler[str, int|str] from the outset.

Outside of that corner case, should be entirely correct, I believe.